fix: rename method
